masaustu_segmentasyon/requirements.txt
PyQt6

Pillow
numpy

opencv-python
scikit-image
torch
torchvision
segment-anything
